feat(Episodes-Component): Move Page to Last Episode Watched and more on Media Page
Refactor EpisodesContainer and MangaChaptersContainer components, remove NavPaginateItems component, and update styles.
- Loading Redone
- Var names changed
